Put the map style in a seperate folder, put the map logic in a seperate folder, added dotted line for walking transport mode, fix the icon transport mode button at the bottom sheet when receiving instruction
